% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-aaa-returns.R
\docType{data}
\name{aaa_returns}
\alias{aaa_returns}
\title{Adaptive Asset Allocation Returns}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 7024 rows and 11 columns.
}
\usage{
aaa_returns
}
\description{
Adaptive Asset Allocation Returns
}
\keyword{data}
