% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_asset_allocation.R
\name{estimate_equal_risk_portf}
\alias{estimate_equal_risk_portf}
\title{Estimate equal risk portfolio weights}
\usage{
estimate_equal_risk_portf(returns, n_days = 60)
}
\arguments{
\item{returns}{An xts object containing returns for two or more assets.}

\item{n_days}{Number of days to use in the covariance matrix calculation.}
}
\value{
An xts object containing weights for each asset in \code{returns}.
}
\description{
This function uses \code{FRAPO::PERC()} to estimate the Equal Risk
Contribution portfolio weights.
}
\author{
Joshua Ulrich
}
