% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pct.R
\name{format_pct}
\alias{format_pct}
\title{Format a number as a percentage}
\usage{
format_pct(x, digits = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{Number of decimal digits}
}
\value{
A character string ending in "\%"
}
\description{
This function formats its input as a percentage.
}
\examples{

format_pct(0.01)  # 1.0\%

}
\author{
Joshua Ulrich
}
