% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_asset_allocation.R
\name{portf_equal_risk}
\alias{portf_equal_risk}
\title{Calculate equal risk portfolio returns}
\usage{
portf_equal_risk(returns, n_days = 120, n_days_vol = 60)
}
\arguments{
\item{returns}{An xts object containing returns for two or more assets.}

\item{n_days}{Number of days of returns to use in the portfolio estimation.}

\item{n_days_vol}{Number of days to use in the covariance matrix calculation.}
}
\value{
An xts object containing the portfolio return for each day in
    \code{returns}.
}
\description{
This function calculates the portfolio weights at the end of each month
using estimated portfolio risk from the returns over the last 60 days.
Then those weights are used to calculate the portfolio returns for the
following month.
}
\author{
Joshua Ulrich
}
