% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_asset_allocation.R
\name{portf_return_momo_min_var}
\alias{portf_return_momo_min_var}
\title{Calculate momentum-ranked minimum variance portfolio returns}
\usage{
portf_return_momo_min_var(
  returns,
  n_assets = 5,
  n_days = 120,
  n_days_vol = 60,
  momo_type = c("relative", "positive", "above average")
)
}
\arguments{
\item{returns}{An xts object containing returns for two or more assets.}

\item{n_assets}{Number of highest momentum assets in the portfolio.}

\item{n_days}{Number of days of returns to use in the portfolio estimation.}

\item{n_days_vol}{Number of days to use in the covariance matrix calculation.}

\item{momo_type}{Type of momentum to use. Can be one of the following:
* relative: use the top \code{n_assets} with the highest momentum.
* positive: use the top \code{n_assets} with the highest *positive*
  momentum. Assets with negative momentum are excluded.
* above average: use the top \code{n_assets} with the highest above-average
  momentum. Assets with below average momentum are excluded.}
}
\value{
An xts object containing the portfolio return for each day in
    \code{returns}.
}
\description{
This function calculates the returns for the global minimum variance
portfolio using the \code{n_assets} with the highest momentum over the last
\code{n_days}.
Then those weights are used to calculate the portfolio returns for the
following month.
}
\author{
Joshua Ulrich
}
