% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_summary.R
\name{portf_summary}
\alias{portf_summary}
\title{Calculate Portfolio Summary Statistics}
\usage{
portf_summary(returns)
}
\arguments{
\item{returns}{An xts object of returns.}
}
\value{
A data.frame of portfolio statistics.
}
\description{
This function takes an \pkg{xts} object of returns and calculates
some portfolio summary statistics.
}
\author{
Joshua Ulrich
}
