% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_asset_allocation.R
\name{portf_top_momentum_min_var}
\alias{portf_top_momentum_min_var}
\title{Calculate momentum-ranked minimum variance portfolio returns}
\usage{
portf_top_momentum_min_var(
  returns,
  n_assets = 5,
  n_days = 120,
  n_days_vol = 60,
  use_momentum = TRUE,
  always_invested = TRUE
)
}
\arguments{
\item{returns}{An xts object containing returns for two or more assets.}

\item{n_assets}{Number of highest momentum assets in the portfolio.}

\item{n_days}{Number of days of returns to use in the portfolio estimation.}

\item{n_days_vol}{Number of days to use in the covariance matrix calculation.}

\item{use_momentum}{Should the portfolio contain only the \code{n_assets}
with the highest momentum (default \code{TRUE}) or all assets?}

\item{always_invested}{Can the portfolio contain assets with negative
momentum (default \code{TRUE}).}
}
\value{
An xts object containing the portfolio return for each day in
    \code{returns}.
}
\description{
This function calculates the returns for the global minimum variance
portfolio using the \code{n_assets} with the highest momentum over the last
\code{n_days}.
Then those weights are used to calculate the portfolio returns for the
following month.
}
\author{
Joshua Ulrich
}
