% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_asset_allocation.R
\name{portf_wts_efficient}
\alias{portf_wts_efficient}
\title{Estimate efficient portfolio weights}
\usage{
portf_wts_efficient(returns, n_days_vol = 20)
}
\arguments{
\item{returns}{An xts object containing returns for two or more assets.}

\item{n_days_vol}{Number of days to use in the covariance matrix calculation.}
}
\value{
An xts object containing weights for each asset in \code{returns}.
}
\description{
This function uses \code{tseries::portfolio.optim} to estimate the
mean-variance efficient portfolio weights.
}
\author{
Joshua Ulrich
}
