% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_asset_allocation.R
\name{portf_return_equal_risk}
\alias{portf_return_equal_risk}
\alias{portf_return_momo_equal_risk}
\alias{portf_return_momo}
\alias{portf_return_momo_min_var}
\alias{portf_return_momo_efficient}
\title{Calculate portfolio returns}
\usage{
portf_return_equal_risk(returns, n_days = 120, n_days_vol = 60)

portf_return_momo_equal_risk(
  returns,
  n_assets = 5,
  n_days = 120,
  n_days_vol = 60,
  momo_type = c("relative", "positive", "above average")
)

portf_return_momo(
  returns,
  n_assets = 5,
  n_days = 120,
  momo_type = c("relative", "positive", "above average")
)

portf_return_momo_min_var(
  returns,
  n_assets = 5,
  n_days = 120,
  n_days_vol = 60,
  momo_type = c("relative", "positive", "above average")
)

portf_return_momo_efficient(
  returns,
  n_assets = 5,
  n_days = 120,
  n_days_vol = 60,
  momo_type = c("relative", "positive", "above average")
)
}
\arguments{
\item{returns}{An xts object containing returns for two or more assets.}

\item{n_days}{Number of days of returns to use in the portfolio estimation.}

\item{n_days_vol}{Number of days to use in the covariance matrix calculation.}

\item{n_assets}{Number of highest momentum assets in the portfolio.}

\item{momo_type}{Type of momentum to use. See Details.}
}
\value{
An xts object containing the portfolio return for each day in
    \code{returns}.
}
\description{
These functions calculate returns for various portfolio specifications. They
calculate portfolio weights at the end of each month and apply those weights
to the following month.
}
\details{
These function use \code{n_days} of returns to estimate the asset weights,
and \code{n_days_vol} to calculate the covariance matrix. Functions with a
\code{n_assets} argument will only include up to that number of assets in
the portfolio.

\describe{
  \item{\code{portf_return_equal_risk}}{uses \code{\link[FRAPO]{PERC}} from the
    \pkg{FRAPO} package to calculate the Equal Risk Contribution portfolio
    return using all assets.
  }
  \item{\code{portf_return_momo}}{calculates the equal weight portfolio of
    the \code{n_assets} with the highest momentum over the past \code{n_days}.
  }
  \item{\code{portf_return_momo_equal_risk}}{calculates the Equal Risk
    Contribution portfolio return using only the \code{n_assets} with the
    highest momentum over the last \code{n_days}.
  }
  \item{\code{portf_return_momo_efficient}}{uses \code{\link[tseries]{portfolio.optim}}
    from the \pkg{tseries} package to calculate the mean-variance efficient
    portfolio return using the \code{n_assets} with the highest momentum over
    the last \code{n_days}.
  }
  \item{\code{portf_return_momo_min_var}}{uses \code{\link[FRAPO]{PGMV}} from
    the \pkg{FRAPO} package to calculate the Global Minimum Variance portfolio
    return using the \code{n_assets} with the highest momentum over the last
    \code{n_days}.
  }
}

The \code{momo_type} argument can be one of the following values:
\describe{
  \item{relative}{use the top \code{n_assets} with the highest momentum.
  }
  \item{positive}{use the top \code{n_assets} with the highest \emph{positive}
    momentum. Assets with negative momentum are excluded.
  }
  \item{above average}{use the top \code{n_assets} with the highest above-average
    momentum. Assets with below average momentum are excluded.
  }
}
}
\author{
Joshua Ulrich
}
