% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_asset_allocation.R
\name{portf_wts_efficient}
\alias{portf_wts_efficient}
\alias{portf_wts_min_var}
\alias{portf_wts_equal_risk}
\title{Estimate portfolio weights}
\usage{
portf_wts_efficient(returns, n_days_vol = 20)

portf_wts_min_var(returns, n_days_vol = 20)

portf_wts_equal_risk(returns, n_days_vol = 60)
}
\arguments{
\item{returns}{An xts object containing returns for two or more assets.}

\item{n_days_vol}{Number of days to use in the covariance matrix calculation.}
}
\value{
An xts object containing weights for each asset in \code{returns}.
}
\description{
These functions estimate weights for various portfolio specifications.
}
\details{
\describe{
   \item{\code{portf_wts_efficient}}{uses \code{\link[tseries]{portfolio.optim}}
       from the \pkg{tseries} package to estimate the mean-variance efficient
       portfolio weights.
   }
   \item{\code{portf_wts_min_var}}{uses \code{\link[FRAPO]{PGMV}} from the
       \pkg{FRAPO} package to estimate the Global Minimum Variance portfolio
       weights.
   }
   \item{\code{portf_wts_equal_risk}}{uses \code{\link[FRAPO]{PERC}} from the
       \pkg{FRAPO} package to estimate the Equal Risk Contribution portfolio
       weights.
   }
}

See the documentation for those respective for details.
}
\author{
Joshua Ulrich
}
