% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_monthly_returns.R
\name{to_monthly_returns}
\alias{to_monthly_returns}
\title{Convert daily returns to monthly}
\usage{
to_monthly_returns(returns)
}
\arguments{
\item{returns}{An xts object of daily returns.}
}
\value{
A series of monthly returns.
}
\description{
This function takes an \pkg{xts} object of daily returns and aggregates
them to a monthly series.
}
\author{
Joshua Ulrich
}
