% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-aaa-returns.R
\name{update_aaa_returns}
\alias{update_aaa_returns}
\title{Update the 'aaa_returns' data set}
\usage{
update_aaa_returns(aaa_returns, src = "tiingo")
}
\arguments{
\item{aaa_returns}{The xts object returned by \code{data(aaa_returns)}}
}
\value{
The \code{aaa_returns} object with returns to the most current date.
}
\description{
This function takes the data set returned by \code{data(aaa_returns)} as input,
pulls data from 2024 to present, calculates returns of the new data, and
appends the new returns to the original data.
}
\author{
Joshua Ulrich
}
