% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_asset_allocation.R
\name{portf_wts_efficient}
\alias{portf_wts_efficient}
\alias{portf_wts_min_var}
\alias{portf_wts_equal_risk}
\title{Estimate portfolio weights}
\usage{
portf_wts_efficient(returns, n_days_vol = 20)

portf_wts_min_var(returns, n_days_vol = 20)

portf_wts_equal_risk(returns, n_days_vol = 60)
}
\arguments{
\item{returns}{An xts object containing returns for two or more assets.}

\item{n_days_vol}{Number of days to use in the covariance matrix calculation.}
}
\value{
An xts object containing weights for each asset in \code{returns}.
}
\description{
These functions estimate weights for various portfolio specifications.
}
\details{
\describe{
\item{\code{portf_wts_efficient}}{estimates the mean-variance efficient
portfolio weights using \code{\link[tseries:portfolio.optim]{tseries::portfolio.optim()}}.
}
\item{\code{portf_wts_min_var}}{estimates the Global Minimum Variance portfolio
weights using \code{\link[FRAPO:PGMV]{FRAPO::PGMV()}}.
}
\item{\code{portf_wts_equal_risk}}{estimates the Equal Risk Contribution
portfolio weights using \code{\link[FRAPO:PERC]{FRAPO::PERC()}}.
}
}

See the documentation for each portfolio weight estimation function for
details on how they work.
}
\author{
Joshua Ulrich
}
